backnews-main/requirements.txt
asgiref==3.8.1
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
distlib==0.3.9
dj-database-url==3.0.0
Django==5.2.1
django-ckeditor==6.7.3
django-js-asset==3.1.2
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docopt==0.6.2
ffmpeg-python==0.2.0
filelock==3.18.0
flutter==0.1
fsspec==2025.5.1
future==1.0.0
gunicorn==23.0.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.5
num2words==0.5.14
packaging==25.0
pillow==11.2.1
platformdirs==4.3.8
psycopg2==2.9.10
PyJWT==2.9.0
regex==2024.11.6
requests==2.32.4
setuptools==80.9.0
sqlparse==0.5.3
sympy==1.14.0
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.31.2
whitenoise==6.9.0
